Rename class from UiState to DataLoadState
